Grouping Wizard events in WizardEventArgs.cs file.
