--- Terminal.Gui/Windows/StepChangeEventArgs.cs
@@ -1,38 +1,0 @@
-﻿using System;
-
-namespace Terminal.Gui {
-
-	public partial class Wizard {
-		/// <summary>
-		/// <see cref="EventArgs"/> for <see cref="WizardStep"/> events.
-		/// </summary>
-		public class StepChangeEventArgs : EventArgs {
-			/// <summary>
-			/// The current (or previous) <see cref="WizardStep"/>.
-			/// </summary>
-			public WizardStep OldStep { get; }
-
-			/// <summary>
-			/// The <see cref="WizardStep"/> the <see cref="Wizard"/> is changing to or has changed to.
-			/// </summary>
-			public WizardStep NewStep { get; }
-
-			/// <summary>
-			/// Event handlers can set to true before returning to cancel the step transition.
-			/// </summary>
-			public bool Cancel { get; set; }
-
-			/// <summary>
-			/// Initializes a new instance of <see cref="StepChangeEventArgs"/>
-			/// </summary>
-			/// <param name="oldStep">The current <see cref="WizardStep"/>.</param>
-			/// <param name="newStep">The new <see cref="WizardStep"/>.</param>
-			public StepChangeEventArgs (WizardStep oldStep, WizardStep newStep)
-			{
-				OldStep = oldStep;
-				NewStep = newStep;
-				Cancel = false;
-			}
-		}
-	}
-}
